Update expected response for `/addsticker`
Recently, Telegram's stickers bot response for `/addsticker` was
changed from "Choose the sticker pack you're interested in." to
"Choose a sticker set." so the userbot was unable to validate the
response as the substring did not match `response.raw_text`. Update
the `/addsticker` expected response string to the latest match.